feat: add recipe form validation and error message

+++ app/forms/recipe_form.rb
@@ -18,11 +18,19 @@
 # ]
 
 class RecipeForm < ApplicationForm
+  with_options unless: -> { @errors.where(:base).present? } do
+    validate :validate_has_at_least_one_gredient,
+             :validate_has_at_least_one_step,
+             :validate_recipe_ingredients
+  end
+
   def initialize(recipe)
     @recipe = recipe
     @cocktail = nil
     @steps = []
+    @steps_size = recipe.steps.size
     @recipe_ingredients = []
+    @recipe_ingredients_size = recipe.recipe_ingredients.size
     @mark_for_destruction_step_ids = []
     @mark_for_destruction_recipe_ingredient_ids = []
   end
@@ -35,11 +43,12 @@ class RecipeForm < ApplicationForm
   end
 
   def save
+    raise ActiveRecord::RecordInvalid unless valid?
+
     ActiveRecord::Base.transaction do
       Step.where(id: @mark_for_destruction_step_ids).destroy_all
       RecipeIngredient.where(id: @mark_for_destruction_recipe_ingredient_ids).destroy_all
       @cocktail&.save!
-      # skip uniqueness validation of position
       @steps.each do |step|
         step.save!(validate: false)
       end
@@ -74,6 +83,7 @@ class RecipeForm < ApplicationForm
           next
         end
       else
+        @recipe_ingredients_size += 1
         recipe_ingredient = @recipe.recipe_ingredients.build
       end
 
@@ -99,10 +109,35 @@ class RecipeForm < ApplicationForm
         end
       else
         step = @recipe.steps.build
+        @steps_size += 1
       end
       step.attributes = { description: step_attrs[:description], position: index }
       index += 1
       step
+    end
+  end
+
+  def validate_has_at_least_one_gredient
+    return if (@recipe_ingredients_size - @mark_for_destruction_recipe_ingredient_ids.size).positive?
+
+    errors.add(:base, I18n.t('activerecord.errors.messages.array_length_too_short',
+                             count: 1, attribute: '原料'))
+  end
+
+  def validate_has_at_least_one_step
+    return if (@steps_size - @mark_for_destruction_step_ids.size).positive?
+
+    errors.add(:base, I18n.t('activerecord.errors.messages.array_length_too_short',
+                             count: 1, attribute: '步驟'))
+  end
+
+  def validate_recipe_ingredients
+    return if @recipe_ingredients.all?(&:valid?)
+
+    @recipe_ingredients.each do |recipe_ingredient|
+      recipe_ingredient.errors.messages.each do |field, messages|
+        messages.each { |message| errors.add(:base, "#{field} #{message}") }
+      end
     end
   end
 end
